extracted toggle function for resolving favor into a module in utils folder
