+++ client/components/SingleFavorView.js
@@ -7,6 +7,8 @@ import { fetchSingleFavor } from "../store/SingleFavor.js";
 import { updateFavor, updateBid } from "../store/favors";
 import useAuth from "./utils/useAuthHook.js";
 import useFavor from "./utils/useFavorHook";
+// experimenting with toggle function in its own module:
+import { toggleFavorResolved } from "./utils/toggleFavorStatus";
 
 const SingleFavor = (props) => {
   // will use this id (from URL param, from route rendering the component)
@@ -24,59 +26,13 @@ const SingleFavor = (props) => {
     dispatch(fetchSingleFavor(props.match.params.id));
   }, []);
 
-  const toggleFavorResolved = async () => {
-    if (favor.status === "CLOSED") {
-      if (favor.bids) {
-        let rejectedBids = favor.bids.filter(
-          (bid) => bid.status === "REJECTED"
-        );
-        let fulfilledBids = favor.bids.filter(
-          (bid) => bid.status === "FULFILLED"
-        );
-        await Promise.all(
-          rejectedBids.map((bid) => {
-            return dispatch(updateBid(bid, { status: "PENDING" }));
-          })
-        );
-        await Promise.all(
-          fulfilledBids.map((bid) => {
-            return dispatch(updateBid(bid, { status: "ACCEPTED" }));
-          })
-        );
-      }
-      await dispatch(updateFavor(favor.id, { status: "OPEN" }));
-      await dispatch(fetchSingleFavor(props.match.params.id));
-    } else {
-      // set the bid(s) with "ACCEPTED" status to status "FULFILLED"
-      // and set all of the bids with "PENDING" status to "REJECTED"
-      if (favor.bids) {
-        let pendingBids = favor.bids.filter((bid) => bid.status === "PENDING");
-        let acceptedBids = favor.bids.filter(
-          (bid) => bid.status === "ACCEPTED"
-        );
-        await Promise.all(
-          pendingBids.map((bid) => {
-            return dispatch(updateBid(bid, { status: "REJECTED" }));
-          })
-        );
-        await Promise.all(
-          acceptedBids.map((bid) => {
-            return dispatch(updateBid(bid, { status: "FULFILLED" }));
-          })
-        );
-      }
-      await dispatch(updateFavor(favor.id, { status: "CLOSED" }));
-      await dispatch(fetchSingleFavor(props.match.params.id));
-    }
-  };
-
   return (
     <div>
       <h1> {favor.title}</h1>
       <br />
       <span> Status: {favor.status === "OPEN" ? "Open" : "Closed"} </span>
       {CurrentUser.id === favor.authorId ? (
-        <button onClick={toggleFavorResolved}>
+        <button onClick={() => toggleFavorResolved(dispatch, favor)}>
           {favor.status === "OPEN" ? "Resolve" : "Reopen"}
         </button>
       ) : (
